requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Limiter==3.8.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
PyJWT==2.9.0
bcrypt==4.2.1
geopy==2.4.1
pandas==2.2.3
redis==5.2.0
python-dotenv==1.0.1
requests==2.32.3
flasgger==0.9.7.1
gunicorn==23.0.0
pytest==8.3.3
pytest-cov==6.0.0
Werkzeug==3.1.3
